requirements.txt
certifi==2019.3.9
chardet==3.0.4
et-xmlfile==1.0.1
graphqlclient==0.2.4
idna==2.8
jdcal==1.4
numpy==1.16.2
openpyxl==2.6.2
pandas==0.24.2
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
urllib3==1.24.2